refactor(admin): :recycle: move edit member to a component
